fixy w formatowaniu

+++ src/objects/resource/ResourceEntity.cpp
@@ -46,7 +46,7 @@ float ResourceEntity::getHealthPercent() const {
 
 Urho3D::String& ResourceEntity::toMultiLineString() {
 	menuString = dbResource->name;
-	menuString += "\nZasobów: " + Urho3D::String(amonut)
+	menuString += "\nZasobów: " + Urho3D::String((int)amonut)
 		+ "\nU¿ytkowników: " + Urho3D::String((int)closeUsers)
 		+ "/" + Urho3D::String((int)maxCloseUsers);
 	return menuString;
